worker/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
gradium>=0.1.0
pydub>=0.25.1
python-multipart>=0.0.20
numpy>=2.0.0
sphn>=0.1.0
python-dotenv>=1.0.0
